[CALCITE-7156] OFFSET and FETCH in EnumerableLimit need to support BIGINT

+++ core/src/main/java/org/apache/calcite/adapter/enumerable/EnumerableLimit.java
@@ -127,7 +127,7 @@ public class EnumerableLimit extends SingleRel implements EnumerableRel {
               Expressions.constant("?" + param.getIndex())),
           Integer.class);
     } else {
-      return Expressions.constant(RexLiteral.intValue(rexNode));
+      return Expressions.constant(RexLiteral.longValue(rexNode));
     }
   }
 }
